backend integration and unit testing of notes-api

+++ notes_api/src/main/java/com/example/demo/repositories/NoteRepository.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.Note;
 
 @Repository
-public interface NoteRepository extends CrudRepository<Note,Integer>{
-	
+public interface NoteRepository extends CrudRepository<Note, Long> {
 
 }
